refactor: simplify hard-coded category list
